# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ClarificationAnswerKind": {
        "description": "How a clarification answer is supplied by the user.",
        "enum": [
          "option",
          "free_text"
        ],
        "title": "ClarificationAnswerKind",
        "type": "string"
      },
      "ClarificationAnswerRequest": {
        "additionalProperties": false,
        "description": "One answer for the currently pending clarification interrupt.",
        "properties": {
          "answer": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "clarification_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Clarification Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ClarificationAnswerKind"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "clarification_id",
          "kind",
          "answer"
        ],
        "title": "ClarificationAnswerRequest",
        "type": "object"
      },
...
      "CreateRunRequest": {
        "additionalProperties": false,
        "description": "User-confirmed text and deterministic interpretation context.",
        "properties": {
          "locale": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Locale",
            "type": "string"
          },
          "reference_date": {
            "format": "date",
            "title": "Reference Date",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source_text": {
            "maxLength": 12000,
            "title": "Source Text",
            "type": "string"
          },
          "timezone": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "source_text",
          "locale",
          "timezone",
          "reference_date"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Versioned error response shared by all future run operations.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PublicError"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "GuestSessionResponse": {
        "additionalProperties": false,
        "description": "Short-lived browser credentials with a bounded server-side owner.",
        "properties": {
          "access_token": {
            "maxLength": 2000,
            "minLength": 32,
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "guest_session_id": {
            "maxLength": 100,
            "minLength": 8,
            "pattern": "^ses_[A-Za-z0-9_-]+$",
            "title": "Guest Session Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "session_expires_at": {
            "format": "date-time",
            "title": "Session Expires At",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "guest_session_id",
          "access_token",
          "token_type",
          "expires_at",
          "session_expires_at"
        ],
        "title": "GuestSessionResponse",
        "type": "object"
      },
...
      "PublicError": {
        "additionalProperties": false,
        "description": "Stable public error without raw provider or stack data.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/PublicErrorCode"
          },
          "details": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "PublicError",
        "type": "object"
      },
      "PublicErrorCode": {
        "description": "Stable error codes that never expose provider or stack details.",
        "enum": [
          "invalid_request",
          "unauthorized",
          "run_not_found",
          "run_conflict",
          "run_expired",
          "rate_limited",
          "provider_unavailable",
          "internal_error",
          "not_implemented"
        ],
        "title": "PublicErrorCode",
        "type": "string"
      },
      "ResumeRunRequest": {
        "additionalProperties": false,
        "description": "Resume from the latest checkpoint after a refresh or interruption.",
        "properties": {
          "last_seen_trace_sequence": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen Trace Sequence"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version"
        ],
        "title": "ResumeRunRequest",
        "type": "object"
      },
...
      "RunResponse": {
        "additionalProperties": false,
        "description": "Versioned envelope shared by create, resume, clarify, and approve.",
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunView"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "run"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "RunView": {
        "additionalProperties": false,
        "description": "A privacy-aware snapshot returned when a run is created or resumed.",
        "properties": {
          "candidate_items": {
            "items": {
              "$ref": "#/components/schemas/ActionItem"
            },
            "title": "Candidate Items",
            "type": "array"
          },
          "clarification_count": {
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Clarification Count",
            "type": "integer"
          },
          "clarification_questions": {
            "items": {
              "$ref": "#/components/schemas/ClarificationQuestion"
            },
            "title": "Clarification Questions",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "run_id": {
            "maxLength": 100,
            "minLength": 8,
            "pattern": "^run_[A-Za-z0-9_-]+$",
            "title": "Run Id",
            "type": "string"
          },
          "safe_trace": {
            "items": {
              "$ref": "#/components/schemas/SafeTraceEvent"
            },
            "title": "Safe Trace",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "schema_version",
          "run_id",
          "status",
          "candidate_items",
          "clarification_questions",
          "clarification_count",
          "safe_trace",
          "created_at",
          "expires_at"
        ],
        "title": "RunView",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SnapFlow API Contract",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/guest-sessions": {
      "post": {
        "description": "Describe anonymous guest creation for generated clients.",
        "operationId": "create_guest_session",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Create Guest Session Contract",
        "tags": [
          "guest sessions"
        ]
      }
    },
    "/api/guest-sessions/refresh": {
      "post": {
        "description": "Describe safe access-token rotation for generated clients.",
        "operationId": "refresh_guest_session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Refresh Guest Session Contract",
        "tags": [
          "guest sessions"
        ]
      }
    },
    "/api/runs": {
      "post": {
        "description": "Describe creation without exposing the route in the runtime app.",
        "operationId": "create_run",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Create run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/clarifications": {
      "post": {
        "description": "Describe one typed clarification answer.",
        "operationId": "answer_clarification",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 8,
              "pattern": "^run_[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Answer clarification",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/resume": {
      "post": {
        "description": "Describe checkpoint resume for the generated client.",
        "operationId": "resume_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 8,
              "pattern": "^run_[A-Za-z0-9_-]+$",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Resume run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}